==> R/00_all.qmd <==
---
title: "00_all.qmd"
format: html
editor: visual
toc: true
embed-resources: true
---

```{r}
#| include: false
library("quarto")
```

# 01_load

{{< include 01_load.qmd >}}

\
\

# 02_clean

{{< include 02_clean.qmd >}}

\
\

# 03_augment

{{< include 03_augment.qmd >}}

\
\

# 04_describe

{{< include 04_describe.qmd >}}

\
\

# 05_antibody_expression_analysis

{{< include 05_antibody_expression_analysis.qmd >}}

\
\

# 06_pca_antibody

{{< include 06_pca_antibody.qmd >}}

\
\

# 07_HLA_allele_analysis

{{< include 07_HLA_allele_analysis.qmd >}}

\
\

# 08_pca_HLA

{{< include 08_mca_HLA.qmd >}}

```{r}
#| include: false
quarto_render("01_load.qmd")
quarto_render("02_clean.qmd")
quarto_render("03_augment.qmd")
quarto_render("04_describe.qmd")
quarto_render("05_antibody_expression_analysis.qmd")
quarto_render("06_pca_antibody.qmd")
quarto_render("07_HLA_allele_analysis.qmd")
quarto_render("08_mca_HLA.qmd")
```

---
title: "00_all.qmd"
format: html
editor: visual
toc: true
embed-resources: true
---
