==> R/03_augment.qmd <==
---
title: "03_augment"
format: html
editor: visual
toc: true
embed-resources: true
---

```{r}
#| include: false
library("tidyverse")
```

### Select data

```{r}
read_tsv("./../data/02_dat_clean.tsv",
         show_col_types = FALSE) |>
  filter(!str_detect(string = sample_name,
                     pattern = "Denver")) |> 
  arrange(case_control, id, timepoint) |> 
  pivot_longer(cols = c(GAD65, ZnT8, IA_2, IAA),
               names_to = "antibody", 
               values_to = "expression") |> 
  pivot_longer(starts_with('HLA_'), 
               names_to = "allele",
               values_to = "subtype") |>
  mutate(allele = str_remove_all(string = allele,
                                 pattern = "(\\,\\d)")) |>
  write_tsv("./../data/03_dat_aug.tsv")
```
