==> R/05_antibody_expression_analysis.qmd <==
---
title: "05_antibody_expression_analysis"
format: html
editor: visual
toc: true
embed-resources: true
---

```{r}
#| include: false
library("tidyverse")
```

### Load Data

```{r}
data <- read_tsv("./../data/03_dat_aug.tsv",
                 show_col_types = FALSE)
```

### Data Analysis

```{r}
#| warning: false
data |>
  group_by(antibody, timepoint, case_control) |>
  summarise(mean_expression = mean(expression, 
                                   na.rm = TRUE),
            .groups = "drop") |>
  ggplot(mapping = aes(x = factor(timepoint),
                       y = mean_expression,
                       fill = case_control)) +
  geom_bar(stat = "identity",
           position = "dodge",
           alpha = 0.8,
           color = 'black') +
  facet_wrap(vars(antibody),
             scales = "free_y",
             nrow = 2) +
  labs(title = "Antibody expression across visits stratified by case and control",
       x = "Timepoint",
       y = "Mean Expression",
       fill = "Case or Control") +
  scale_fill_manual(values = c(Case = "#481567",
                               Control = "#20A387")) +
  theme_bw() + 
  theme(legend.position = "bottom")

ggsave(width = 8,
       height = 6,
       filename = "05_plot_1.png",
       path = "./../results/")
```
