==> R/06_pca_antibody.qmd <==
---
title: "06_pca_antibody"
format: html
editor: visual
toc: true
embed-resources: true
---

```{r}
#| include: false
library("tidyverse")
library("broom")
library("ggrepel")
```

### PCA handling and plotting

```{r}
conversion_table <- tibble(
  case_control = c("Case", "Control"),
  num_case_control = c(1, 0)
)

data <- read_tsv("./../data/03_dat_aug.tsv",
                 show_col_types = FALSE) |>
  pivot_wider(names_from = "antibody",
              values_from = "expression") |>
  select(case_control, GAD65, IA_2, IAA, ZnT8) |>
  mutate(across(starts_with("GAD65"), 
                ~ if_else(is.na(.), 0.001, .)), 
         across(starts_with("IA_2"), 
                ~ if_else(is.na(.), 0.001, .)), 
         across(starts_with("IAA"), 
                ~ if_else(is.na(.), 0.001, .)), 
         across(starts_with("ZnT8"), 
                ~ if_else(is.na(.), 0.001, .))) |> 
  left_join(conversion_table, 
            by = "case_control")

data_pca <- data |>
  select(-case_control,
         -num_case_control)|>
  scale() |>
  prcomp()

data_pca |>
  augment(data) |>
  ggplot(data = _,
         mapping = aes(x = .fittedPC1, 
                       y = .fittedPC2,
                       color = case_control)) + 
  geom_point(size = 1.5) +
  scale_color_manual(values = c(Case = "#481567",
                                Control = "#20A387")) +
  theme_bw() +
  labs(title = "Plot of individuals",
       color = "Case vs Control")

ggsave(width = 8,
       height = 5,
       filename = "06_plot_1.png",
       path = "./../results/")

data_pca |>
  tidy(matrix = "eigenvalues") |>
  ggplot(mapping = aes(x = PC,
                       y = percent)) +
  geom_col(fill = "#20A387", 
           alpha = 0.8) + 
  scale_x_continuous(breaks = 1:9) +
  scale_y_continuous(labels = scales::percent_format(),
                     expand = expansion(mult = c(0, 0.01))) +
  geom_line(mapping = aes(x = PC),
            color = "#481567") +
  theme_bw() +
  labs(title = "Explained Variance by Principal Components",
       y = "Percentage Contribution")

ggsave(width = 8,
       height = 5,
       filename = "06_plot_2.png",
       path = "./../results/")

arrow_style <- arrow(length = unit(0.2, "cm"), 
                     ends = "first")

num_case_control_coord <- cor(data$num_case_control, data_pca$x) |>
  as_tibble()

data_pca |>
  tidy(matrix = "rotation") |>
  pivot_wider(names_from = "PC", 
              names_prefix = "PC", 
              values_from = "value") |>
  ggplot(mapping = aes(x = PC2,
                       y = PC1)) +
  geom_point(alpha = 0) + 
  coord_flip() +
  theme_bw() +
  labs(title = "Rotation Matrix",
       subtitle = "Exploring Variability Along Principal Components") +
  geom_segment(mapping = aes(xend = 0,
                             yend = 0),
               arrow = arrow_style,
               color = "#481567") +
  geom_segment(data = num_case_control_coord,
               mapping = aes(xend = 0,
                             yend = 0),
               arrow = arrow_style,
               color = "#20A387") +
  geom_label_repel(data = num_case_control_coord,
                  mapping = aes(label = "case_control"),
                  color = "#20A387",
                  size = 3.5) +
  geom_label_repel(mapping = aes(label = column),
                  color = "#481567",
                  size = 3.5)

ggsave(width = 8,
       height = 5,
       filename = "06_plot_3.png",
       path = "./../results/")
```
