==> R/07_HLA_allele_analysis.qmd <==
---
title: "07_HLA_allele_analysis"
format: html
editor: visual
toc: true
embed-resources: true
---

```{r}
#| include: false
library("tidyverse")
```

```{r, fig.height=10, fig.width=8}
#| warning: false
data <- read_tsv("./../data/03_dat_aug.tsv",
                 show_col_types = FALSE)

HLA_tidy <- data |> 
  filter(timepoint == 4) |> 
  mutate(allele = str_remove(string = allele,
                             pattern = "(\\,1)|(\\,2)")) |>
  mutate(allele_subtype = str_c(allele,
                                subtype,
                                sep = ":")) |> 
  drop_na(allele_subtype) |> 
  mutate(subtype_n = 1) |>
  select(case_control, id, allele_subtype, subtype_n) |> 
  group_by(case_control, allele_subtype) |>
  summarize(allele_n = n(), .groups = "drop") |> 
  pivot_wider(names_from = case_control,
              values_from = allele_n) |> 
  na.omit(Control, Case) |> 
  mutate(ratio = round(Case/Control, 
                       digits = 2)) |>
  select(allele_subtype, ratio)

ggplot(data = HLA_tidy,
       mapping = aes(x = reorder(allele_subtype, +ratio),
                     y = ratio)) + 
  geom_bar(stat = "identity",
           color = "#20A387",
           fill = "#20A387",
           alpha = 0.8) + 
  coord_flip() + 
  labs(title = "Ratio of HLA subtype occurence for case and control patients",
       x = "Allele subtype",
       y = "Ratio (Case/Control)") + 
  theme_bw() + 
  theme(panel.grid.major.y = element_blank(),
        axis.text.x= element_text(size=10),
        axis.text.y= element_text(size=7))

ggsave(width = 10,
       height = 10,
       filename = "07_plot_2.png",
       path = "./../results/")
```

```{r, fig.height=6, fig.width=12}
#| warning: false
#| message: false
data <- read_tsv("./../data/03_dat_aug.tsv",
                 show_col_types = FALSE)

HLA_tidy <-
  HLA_tidy |>
  ungroup() |>
  mutate(allele = str_extract(string = allele_subtype,
                              pattern = "HLA[^\\:]+"),
         subtype = str_extract(string = allele_subtype, 
                               pattern = "\\d{4}"))


ggplot(data = HLA_tidy,
       mapping = aes(x = subtype,
                     y = allele)) +
  geom_tile(aes(fill = ratio),
            color="#481567") + 
  theme_bw()+
  theme(axis.text.y = element_text(size = 10),
        axis.text.x = element_text(angle = 90,
                                   size = 10,
                                   vjust=0.5),
        panel.grid.major.y = element_blank())+
  scale_fill_gradient(low = "#20A387",
                      high = "#481567") + 
  labs(title = "Ratio of HLA subtypes in case and control patients",
       x = "Subtype",
       y = "Allele",
       fill = "Case/Control")

ggsave(width = 8,
       height = 5,
       filename = "07_plot_1.png",
       path = "./../results/")
```

\
