==> R/08_mca_HLA.qmd <==
---
title: "08_mca_HLA"
format: html
editor: visual
toc: true
embed-resources: true
---

```{r}
#| include: false
#| message: false
library("tidyverse")
library("FactoMineR")
library("ggrepel")
library("purrr")
```

## Load Data

```{r}
#| message: false
data <- read_tsv("./../data/03_dat_aug.tsv", 
                 show_col_types = FALSE)
```

## Analysis

### MCA HLA subtypes

#### Percentage of variance explained

```{r}
HLA_tidy <- data |> 
  filter(timepoint == 1) |>
  select(case_control, id, allele, subtype) |>
  mutate(allele_subtype = str_c(allele,
                                subtype,
                                sep = ":")) |>
  na.omit(allele_subtype) |>
  group_by(case_control, id, allele_subtype) |>
  summarize(subtype_n = n(),
            .groups = "drop") |>
  mutate(subtype_n = str_replace(string = subtype_n,
                                 pattern = "4", "yes")) |>
  pivot_wider(names_from = allele_subtype,
              values_from = subtype_n) |>
  mutate(across(everything(),
                ~replace_na(.x, "no"))) |>
  mutate(across(starts_with("HLA"), as.factor)) |>
  mutate(patient = str_c(case_control,
                         id,
                         sep = ":"))



relevant_var <- unname(as_vector( HLA_tidy |>
                                    pivot_longer(cols = starts_with("HLA"),
                                                 names_to = "subtype",
                                                 values_to = "presence") |>
                                    select(subtype, 
                                           presence) |>
                                    filter(presence == "yes") |>
                                    group_by(subtype) |>
                                    summarize(freq = n()/54) |>
                                    ungroup() |>
                                    filter(freq >= 0.30) |>
                                    select(subtype) ))



HLA.active <- HLA_tidy |>
  select(all_of( relevant_var ))

mca_fit <- MCA(HLA.active, 
               graph = FALSE)



as_tibble_row(mca_fit) |>
  select(eig) |>
  unnest(eig) |>
  reduce("*") |>
  as_tibble() |>
  rename("eigenvalue" = 1,
         "var_percent" = 2,
         "cum_var_percent" = 3) |>
  mutate(Dim = str_c("Dim", seq(1,13)))|>
  ggplot(aes(x = reorder(Dim, - var_percent),
             y = var_percent)) +
  geom_col(fill = "#481567",
           alpha = 0.8) +
  geom_text(aes(label = round(var_percent,1)),
            size = 3,
            vjust = -0.3) +
  theme(axis.text.x = element_text(angle = 90),
        panel.grid.minor = element_blank()) +
  theme_bw() +
  labs(title = "Variance Explained by Each Dimension",
       x = "Dimension",
       y = "Percentage of Variance explained")

ggsave(width = 8,
       height = 5,
       filename = "08_plot_1.png",
       path = "./../results/")

```

#### Allele subtype category (present or not) representation

```{r}
coordinates <- as_tibble_row(mca_fit) |>
  select(var) |>
  unnest_wider(var) |>
  select(coord)|>
  unnest(coord) |>
  reduce("*") |>
  as_tibble() |>
  select(1,2) |>
  rename("coord_Dim1" = 1,
         "coord_Dim2" = 2) |>
  mutate(id=seq(1,26))
  
cos2values <- as_tibble_row(mca_fit) |>
  select(var) |>
  unnest_wider(var) |>
  select(cos2)|>
  unnest(cos2) |>
  reduce("*") |>
  as_tibble() |>
  select(1, 2) |>
  rename("cos2_Dim1" = 1,
         "cos2_Dim2" = 2) |>
  mutate(id = seq(1,26))


full_join(coordinates,
          cos2values,
          by = join_by(id == id)) |>
  mutate(HLA = rownames(data.frame( mca_fit$var )),
         representation = cos2_Dim1 + cos2_Dim2) |>
  ggplot(data = _,
         mapping=aes(x = coord_Dim1,
                     y = coord_Dim2,
                     label = HLA)) +
  geom_vline(xintercept = 0,
             linetype = "dashed",
             linewidth = 0.2) +
  geom_hline(yintercept = 0,
             linetype = "dashed",
             linewidth = 0.2) +
  geom_point(aes(color = representation)) +
  geom_label_repel(aes(color = representation),
                   force_pull = 0,
                   size = 2,
                   nudge_y = 0.4,
                   direction = "y") +
  scale_color_gradient2(low="#20A387",
                        mid="#f68f46ff",
                        high="#481567",
                        midpoint=0.4) +
  theme_bw() +
  labs(title = "Dimension 2 vs. Dimension 1",
       subtitle = "Allele subtype Categories stratified by dimensional representation of category",
       x = "Dim 1",
       y = "Dim 2",
       color = "Cos2")

ggsave(width = 8,
       height = 5,
       filename = "08_plot_2.png",
       path = "./../results/")
```

#### Individuals in 2D.

```{r}
as_tibble_row(mca_fit) |>
  select(ind) |>
  unnest_wider(ind) |>
  select(coord)|>
  unnest(coord) |>
  reduce("*") |>
  as_tibble() |>
  select(1, 2) |>
  rename("coord_Dim1" = 1,
         "coord_Dim2" = 2) |>
  mutate(case_control = HLA_tidy |> pull(case_control),
         patient = HLA_tidy |> pull(patient)) |>
  ggplot(data = _,
       mapping = aes(x = coord_Dim1,
                     y = coord_Dim2,
                     label = patient)) +
  geom_vline(xintercept = 0,
             linetype = "dashed",
             linewidth = 0.2) +
  geom_hline(yintercept = 0,
             linetype = "dashed",
             linewidth = 0.2) +
  geom_label_repel(mapping = aes(color = case_control),
                   size = 2,
                   max.overlaps = 15) +
  scale_color_manual(values = c(Case = "#481567",
                                Control = "#20A387")) +
  theme_bw() +
  labs(title = "Dimension 2 vs. Dimension 1",
       subtitle = "Patients stratified by outcome",
       x = "Dim 1",
       y = "Dim 2",
       color = "Case/Control")

ggsave(width = 8,
       height = 5,
       filename = "08_plot_3.png",
       path = "./../results/")
```

None of the dimensions explain enough of the variance themselves. The allele subtype category representation of dimension 1 and 2 show promise of investigative allele subtypes, however since the variance explained by the dimensions is low, no conclusion can be drawn. The patient representation gives less evidence of sufficient variance explained by dimension 1 and 2. A four dimensional explanation of variance could be investigated, as the first 4 dimensions explain a little over 70% of the variance.
